#1486 Improve ArticlePage layout
